Give Exit Full Screen and Configure buttons text and adjust wording
Summary:
Unlike Kate, Gwenview's {nav Exit Full Screen} button is icons-only, impeding
discoverability. This patch shows its action's text, and adjusts that text to be more
consistent with the typical wording.

In the process, the button below it is also given visible text to avoid making it look
odd in its layout.

To accomplish this technically, the `FullScreenToolBar` class's `addAction()` function
is given a new optional parameter for specifying if and where the button's text should
appear, which is then used to give the aforementioned buttons visible text according to
the user's toolbutton text display preferences. The actions are also no longer marked
with the Low Priority status, which forces them to be always icons-only and was
previously redundant since icons-only status was forced in a different way. Fixed now.

BUG: 385314
FIXED-IN: 20.04.0

Test Plan: {F8176526}

Reviewers: #gwenview, #vdg, broulik

Reviewed By: broulik

Tags: #gwenview

Differential Revision: https://phabricator.kde.org/D28052