Improve responsiveness when showing single image without thumbnail bar
Editing an image in a huge folder before this patch is pretty
frustrating because there's a ton of (unnecessary) stuff happening in
the background.

In this case it is the thumbnail delegates that constantly fetch and
convert qmodelindexes to positions to check if they need to update their
hover ui, even if the hover ui things are off. Further improvements
would be to actually check if the hover stuff is used, but with this
patch things seem pretty smooth at least.

Differential Revision: https://phabricator.kde.org/D28373