Fix hang on exit when viewing huge folder
Even if thumbnail storing is turned off in the settings gwenview still
stores them and instead deletes them on exit. I have a separate fix for
that, but this also meant that it hung for a long time when exiting
because it didn't try to shut down the thumbnail writer thread.

Differential Revision: https://phabricator.kde.org/D28374