Make it possible to shut up gwenview with e.g. kdebugdialog
exiv2 lets us register a log handler, so it's fairly straight forward to
put our own that maps it to qC{Warning,Debug,Info} with a proper
category. This allows people who're not interested in exiv2 spam turn it
off.

It only shows "Critical" by default, so supress everything, the warning
messages aren't really useful and the "real" error messages are already
stored in Exiv2ImageLoader.

Differential Revision: https://phabricator.kde.org/D28376