Fix ESC key not exiting fullscreen when viewing image
The event filter was not catching the QEvent::ShortcutOverride
event in browsemainpage.cpp and therefore when the ESC
key was pressed, gwenview would browse images instead
of leaving fullscreen. This commit ensures that when
the ESC key is pressed the first time, gwenview will
exit fullscreen and will browse images on second key
press.

BUG: 425479