Fix window title not being shown by embedded KPart
DocumentViewPrivate::updateCaption() will not change the caption when
mCurrent is false (as introduced by commit a7819f6e). It is correctly
set when using the Gwenview application, but it is never set by the part
and so the window title is never set.

Since there is no image selection or slideshow when using the part,
it is enough to simply call setCurrent() in openUrl().