Do not restore the JPG quality after a "Save" operation
The "Save as" dialog allows to save images with different JPG quality.
After performing the "Save as" operation the JPEG quality was restored
to the original gwenview configuration.

However if the user only saved the modified image this same restoration
process was performed too. But in this case the variable
(configFileJPEGQualityValue) had the value which have been set at startup
not on the start of the save process, which lead to discarding the user
changes made to the JPG save quality since the program start.

Moved the restoration process to be ran only for the "Save as" operation.

Fixes: 423414