Fix zoom with F or middle mouse click caused by the introduction of the spinbox in 2cdaca95
With the introduction of the spin box, the F button / middle mouse click don't work properly anymore.
Problem caused because the spinbox did an emitZoomChanged even tho the zoom was changed by something else than the spinbox. In the case from the bug report it is 'Fit to window'.
Spin box emits now only after it had focus - either by mouse scroll change or manually typing a value into it.

BUG: 432075