Show quality chooser for other lossy image formats too
WEBP, AVIF, HEIF, and HEIC are also compressed image formats, so it
makes sense to allow the user to choose the compression level for them
too, just like with JPEG images.

This commit makes the internal variable names a little odd since they
still have "jpeg" in them. This can be corrected in a follow-up commit.

BUG: 431680
FIXED-IN: 21.04