Only generate thumbnails for items that are visible or may soon be visible
Rather than generating thumbnails for everything, which will lead to a
lot of invisible thumbnails being kept in memory, easily ballooning
memory usage to large amounts for a directory with 10000 items.

An attempt is made to keep some additional thumbnails next to the ones
currently visible so when scrolling they are already loaded, but that
process is not exhaustive.