Only navigate to the first document if we do not have a current URL
If we pass in a URL from the command line, there's a chance it will not
be selected by the time the dir lister finished. If that happens, we are
resetting the view to the first item, which hides the image we were
currently viewing. Instead, only navigate to the first item if we do not
have a current URL so we do not change the current URL.