Use a simple QGraphicsItem subclass for drawing raster images
RasterImageView was using a custom implementation that would use two
QPixmaps to do double buffered rendering. However, since Qt already
renders things double buffered, we are doing duplicate work. So drop
that and use a simple QGraphicsItem subclass that paints the loaded
image directly. This simplifies things a lot and removes a fair amount
of difficult to follow custom code.

BUG: 271671