Hide the side bar by default
When a user opens an image they are primarily interested in
viewing the image, not in editing tools, file operations or meta
data. Hiding the side bar in view mode leads to a nicer default
viewing experience.

The editing tools of the side bar are already more easily
discoverable than they used to be because of the new button in the
default toolbar for showing them.

In browse mode on the other hand everything in the "Information"
sidebar tab can also be accessed through the "Properties" action
both in the "Operations" sidebar tab or in the right-click context
menu. So users should be able to edit the meta data even if they
won't see by default that an "Information" sidebar tab even exists.
With the "Show Editing Tools" toggle action in the toolbar and the
menu bar or hamburger menu they should be able to find whatever
else they are looking for.

Having too many panels and bars visible by default doesn't
automatically lead to better discoverability of all the
functionality because users will start to tunnel vision on the
image they are viewing instead of trying to make sense of all the
UI elements.