Render raster images by using a copy of a portion of the image
Instead of directly rendering the image, we now copy the visible area of
the image and render that. This allows us to perform color correction on
the copy which does not impact the original data and thus should
preserve the original pixels. As a bonus, this also allows us to use
QImage's improved smooth scaling so reduces aliasing artifacts at small
zoom values.