Wrap around during navigation with arrow keys
This resurrects https://phabricator.kde.org/D9039 originally by Nate Grahm. By default when on last image pressing the right arrow will wrap to the first image, and left arrow on first image will wrap to the last. to determine the proper next and last images, it skips things that are not images