Conserve sidebar visibility between browse and view mode
Before this commit, the visibility of the sidebar was separately
saved in browse mode and in view mode. This lead to the irritating
situation that the sidebar could somewhat unexpectedly show or hide
itself when switching between browse and view mode.

With this commit the sidebar will only be set visible when the user
explicitly showed it and the sidebar will only be set invisible
when the user explicitly hid it.

The fullscreen mode is exempt from this because having the sidebar
visible in fullscreen seems to go against why someone would trigger
the fullscreen mode in the first place.