Use decorated tags in info sidebar
It would probably be a good idea to turn these into interactive tags one day. For now, they are not interactive, but at least they make the list of tags more attractive and readable. If they are ever made interactive, they should probably be changed to have a more interactive looking appearance.