Remove Spacebar and Backspace as navigation keys in View mode
These keys are problematic for a few reasons:
1. Spacebar conflicts with play/pause, so when you navigate to a video,
   you can't navigate away from it using the spacebar key anymore
2. The Backspace key is a slightly scary key to use for navigation since
   in all other contexts, it means "delete this thing"

In the past, the arrow keys did not trigger wrap or the "you've reach
the beginning/end" dialog, so you had to use space/backspace to do those
things. However this has since been corrected. So we can safely use only
the arrow keys for navigation.