Hide the menu bar by default
This commit hides the menu bar for new installations by default.
This is done by introducing a setting to save if Gwenview was
ever started before. Unfortunately this check doesn't work for
upgrading users so the menu bar is hidden only if a hamburger menu
action is currently in the toolbar which can only be the case if
it is a new installation.

Motivation:
With the hamburger menu now visible on the toolbar by default, we
can now hide the menu bar by default without having to fear that
users won't be able to access Gwenview's features.

Hiding the menu bar by default makes sense because Gwenview is
supposed to be a simple image viewer meaning that users shouldn't
need fast access to the plethora of actions in the menu bar.

Hiding the menu bar by default increases the default area available
for showing an image and directs users to be instead more aware of
the select view actions which remain visible on the UI.