Fix licencing situation
Gwenview is mostly licenced under GPL 2.0 or later. Some
contributors in the last month (including me) were using LGPL
licences that are more restrictive about the specific version.
This commit changes licences with approval of the copyright
holder so that the licencing situation is kept the same way as it
was orignally intended by prior maintainers.