Align the side bar on the first run by setting its width
To make Gwenview pretty the sidebar is supposed to align with a
separator in the toolbar. For narrow side bar widths this is not
possible because the separator is too far away from the window
edge.

This commit changes it so the side bar width is set to a nice
default value on the first run.