Allow typing a zoom level in the zoom combo
This fixes a bug that would sometimes disallow typing a zoom level
in the zoom combobox (in the statusbar).

For instance if the user has 400% zoom and wants to manually set
it to 380%, they would select the full text and start typing.
When the first 3 is typed there is a chance that the combobox would
replace the typed '3' with a much higher value based on the mininum
allowed. For instance '13'.
The effect is a 4 digit zoom instead of the one the user wanted.

This commit allows there to be an invalid state which we do not
act on (zoom does not change) while the user is typing.