Don't change the zoom value unless the user made a selection
This is a fix for a feature that was recently introduced.

The ZoomComboBox in the very bottom right of the Gwenview main
window can be used to select different zoom values. The zoom values
are changed when they are hovered so users have a way to know
which zoom these values correspond to in relation to the size of
the currently viewed image.

Unfortunately when a user hovered any option and then closed
the popup without making a selection, the zoom value wasn't
correctly reset to the zoom value that was used before the popup
was opened. This commit fixes this.

This commit also makes it so when the popup is opened, the option
that corresponds to the current zoom value is highlighted. If the
current zoom value doesn't exist as an option in the popup,
nothing is highlighted by default.

BUG: 441004
FIXED-IN: 21.12