Fix scrolling on the ZoomComboBox
Scrolling on the ZoomComboBox should go through the list of
predefined zoom values of the box. Before this patch this
was very unreliable. For example it didn't work as expected when
the current zoom value wasn't one of the predefined values.

With this commit, scrolling should work exactly as expected.

There is one change that also affects people that aren't scrolling:
The smallest zoom value is no longer generated in the list of
predefined zoom values because this value is already covered by
the zoom value "fit" (and by "100%" if the image is very small).