Add explicit WITH_X11 CMake option
HAVE_X11 already exists but is set automagically so far, but using
-DCMAKE_DISABLE_FIND_PACKAGE_X11 will break if any dependencies list X11
as required in their cmake config.

We want to be able to build without X11 support even if some of the used
libraries may not work w/o X11 themselves yet or need to be built with
X11 support for other reverse dependencies.

Keep the option confined to UNIX AND NOT APPLE.

Signed-off-by: Andreas Sturmlechner <asturm@gentoo.org>