Only switch to Qt::FastTransformation ("showing pixels") at 500% zoom
Right now Gwenview uses Qt::FastTransformation to show individual pixels
at any zoom value higher than 100%. This is not optimal and results in
poor display for low zoom levels where a smoothed image is more
aesthetically desirable. But there are use cases for seeing individual
pixels--for example when looking at the details of a line-art graphic
or an icon, or examing an image to see its individual colors. These use
cases becomes more likely as the zoom level is increased.

Accordingly, this commit changes Gwenview to shop smoothing zoomed-in
images at an arbitrarily-chosen threshold of 400% zoom. At that level or
above, the image will be unsmoothed and you can see the individual
hard-edged pixels.

BUG: 443010
FIXED-IN: 22.04