Remove background color chooser widget
In e6241375e3ffecf0bfc0973b5e00dd193183eabb actions were added to
Gwenview which allowed easy changing of the background color. As
part of this change a widget was added to the status bar which
contains these new actions.

This commit removes the widget again while keeping the actions to
quickly change the background color of the view available both from
the menu bar and from the right-click context menu.

Having these actions always visible on the status bar doesn't seem
necessary because users generally seem to not be too interested in
switching the background color frequently.

This commit is kind of a requirement for adding other actions to
the status bar instead which users are vocally more interested in
having.

Most of the added lines were just moved from the widget class to the `documentviewcontroller`.

CCBUG: 441447