Revert "Prefer mime type from content over file name when loading"
This reverts commit 6a79391a9ab68bec839897369c665ee2e3afe7e5.

That commit caused Gwenview to lose the ability to open RAW files. This
is a more serious regression than the bug being fixed by the commit
(329140), so it needs to be reverted for now until that bug can be
fixed in a way that doesn't regress RAW loading.

BUG: 441698
FIXED-IN: 21.12.2

CCMAIL: ahiemstra@heimr.nl