Prevent users from "losing" the thumbnail bar
The thumbnail bar can be freely resized by users. It is quite easy
to resize it in a way that collapses it to a size of zero which
makes the bar invisible. Prior to this commit, the only way to get
it back is by finding the invisble drag handler and resizing it
to a non-zero size. Not even an application restart helps. This
isn't intuitive at all.

With this commit, the bar being invisible because of a resize is
treated similarly to the bar being hidden through a button press.
This means that the "Show Thumbnails" button always works in
toggling the visibility of the bar even if it has been resized to 0
by a user prior to that.

I decided to not do the easy fix of disallowing users to resize the
bar to 0 because it is a very intuitive way of hiding the bar and
preventing users from hiding it this way seems weird to me.