Add tooltips for "Fit" and "Show Thumbnails" actions
"Fit" is a bit non-descriptive so a tooltip is added: "Fit image into the
viewing area".

Its text is changed from "Zoom to fit" to "Zoom to Fit" (capital "F") to
conform with the HIG.

An expanded tooltip is also added so users who might not understand why
smaller images aren't enlarged are aware of the setting "Enlarge smaller
images". A hyperlink is added to the help text that will open the
relevant settings page.

"Show Thumbnails" also gets an expanded tooltip mostly to make
users aware of the cool functionalities and settings it has.

Some other useful information is also added to the help texts.