Add global scope left/right arrow key shortcuts for the view
The thumbnails list responds to the left and right arrow key, but if the
view happens to be fixed (e.g. by having clicked on it) then these
shortcuts stop working, because the view doesn't natively handle the
left and right arrow keys for this, and the applicable actions in the
actionCollection don't have shortcuts bound.

This commit adds those shortcuts. They don't conflict with anything else
that uses it because any local handling of these keys (e.g in the
thumbnail strip) will override them.