Add a splitter to the information sidebar
In Gwenview, the information page of the sidebar consists of two halves: an upper section displaying varying amounts of image details,
and a lower section displaying a rating, tags, and description.
Due to the expanding nature of the description's plain text edit, it takes up half of the sidebar by default, and is usually completely empty.
This additional space could be better used to show more image details, as requested in this bug report:

BUG: 459478

So, I have added a splitter, allowing you to move the previously-static central divider, 
and made sure that Gwenview remembers its position between sessions.

By default, the appearance of the sidebar looks identical to how it did before this patch.

Below are two screenshots, one of the splitter moved to its highest, and another at its lowest position.

![split_high](/uploads/1996ac1cd5be7d22cee2dacd6ca8ce7c/split_high.png)

![split_low](/uploads/a040f1d115efab6d7bc6a2f3b1489dc6/split_low.png)