Save a shallow copy of the original image when rotating
BUG: 439697
FIXED-IN: 23.04

The app crashes when trying to get an image from a document using
document->image(). Looks like it can be destroyed from another thread
so we should make a shallow copy of it. It's fast and fixes the crash.


(cherry picked from commit 5ddac2cc9ff426302771fb6c7aceae02953b945e)