Remove "Delete" and "Create Folder" actions from sidebar
The number of actions in Gwenview's sidebar currently causes the main
window with default settings to be taller than the amount of vertical
space available for users with 1366x768 screens and the default 44px
bottom panel. In this configuration, Gwenview's footer becomes
inaccessible, breaking various parts of the app's functionality, such as
the ability to zoom and switch sidebar tabs.

I know what you're about to say. Yes, it's 2023 and people are still
using 1366x768 screens. Unfortunately they still will be for a long
time; this is the second most common screen resolution for
desktop/laptop computers worldwide. See
https://gs.statcounter.com/screen-resolution-stats/desktop/worldwide

We could fix this by making the sidebar actions list scrollable, but
that's kind of an ugly solution as scrollable lists of actions
present an unpleasant UX overall.

So instead, let's remove two actions from the sidebar:
- The "Delete" action, which is actively dangerous to the user's data
  and should definitely not live next to safe actions.
- The "Create Folder" action is super niche in purpose as Gwenview is
  not really meant to be a file  management app.

Both actions remain visible in the GUI menu structure and accessible
via keyboard shortcut.

BUG: 458987
FIXED-IN: 23.08