Add an action to open in a new window
After Bug 462088 was resolved by removing Gwenview from its own "Open With"
menu, it turned out that some people were using this for various workflows.

To support them in a way that doesn't look like a bug to users without
those workflows, this commit adds an action that allows the user to open
the currently selected item or currently viewed image in a new window.

BUG: 469179
FIXED-IN: 23.08