Delay showing loading indicator
Before this commit, we always instantly showed a loading indicator
as soon as an image started loading. For fast loading images this
meant that we showed a cogwheel loading indicator in the middle of
the screen for a very brief period of time. This loading indicator
was placed on top of other images during a smooth transition and
didn't serve any purpose whenever loading is pretty much instant.

This commit makes it so the loading indicator only shows up after a
split second, thus avoiding unnecessary screen movement.