Make image switching seamless when animations are disabled
Previously, switching images meant that between hiding of the first
image and showing of the next image there would be a blank screen
when animations are disabled. This brief flickering isn't aesthetic
and makes direct comparisons of similar images more difficult.

This commit makes the switching of images seamless when animations
are disabled. It does so by disabling the updating/repainting of
the view container until the loading of the new image has
completed().

If loading takes long enough that we would want to show a loading
indicator the seamless experience is given up on and we do show a
loading indicator instead.

BUG: 472351


(cherry picked from commit 35a5a93867f9f4a5d4bed896df1cb4180cfdac87)