Stop overriding settings menu
In the past this was done to not have the Full Screen action appear
there. However this was changed globally three years ago in
https://invent.kde.org/frameworks/kxmlgui/-/commit/93455f78d78d2002ad0d0044036ed13cfd90fbd4
and as such we haven't needed to do it anymore in Gwenview since then.

BUG: 473672
FIXED-IN: 23.08.1


(cherry picked from commit c63b26c1d8855eafd1374b7856d0ad4edf4589a2)