Rate limit how often we update PreviewItemDelegate when model rows change
Updating the PreviewItemDelegate will trigger a relayout of the view,
which is an expensive blocking operation if the directory the view is
showing is large. Since ThumbnailView::rowsInsertedSignal may be called
quite often while we're loading a directory, we would end up relayouting
way too often. This slows down directory loading and generally reduces
performance. So instead apply it only before painting