thumbnailview: Check if thumbnail is actually visible when generating
visualRect() actually generates a rect that is valid for all items, not
just visible ones. So check if it intersects the viewport to verify that
it's actually visible. This fixes the code that was trying to limit
which thumbnails were being generated.

(cherry picked from commit 315be918c9ef709b09c9bc3e475d6580a3da51e4)