Add minimal view mode
The minimum view mode is a window with as little widgets beside the
image itself as possible.

Add a mode switch (minimum view) when activated, hide all widgets, add
an exit button. Save the necessary settings for the state of widgets
and view mode - the next time you start the application, they will be
restored.