Adds a new filter ImageDimensionsFilter allowing filtering the images by their width and height.
Adds a new filter `ImageDimensionFilter` to filter the images by their width and height. Allows to filter the images with the `GreatherThan`, `Equals` and `LessThan` comparators.

This filter is added in the combox box that contains the filters. Once selected, a new combo box is added to select the comparator (like in `DateFilter`) and two QLineEdit are added with a corresponding QLabel where user can introduce the image size in pixels.

@teams/usability @ahiemstra