Make Name Filters Search Recursively
Currently the name filter opts not to handle directories. This means that when a user tries to search for images that contain a certain string. All folders will be shown regardless of relevance. This recursively searches for the user input string in any sub directories contained in the current directory.

I have tested to ensure that there was no substantial memory usage.
I even created a directory with 810 sub directories. With the matching image being in the very bottom directory. Why 810 directories? Either the integrated terminal or fish shell seemed to start having issues when I originally told it to create a thousand nested directories. Anyways despite the large number of sub directories which no real world user should hopefully ever encounter. It still seemed to work extremely fast. At least on my machine, but I am welcome to feedback.

I know that when it comes to these recursive kinds of problems there are times when you should use stack memory and other times when you should use heap memory. Since it seemed to work well even with the crazy number of sub directories, I didn't bother to do anything special with memory allocation. Furthermore, I recall reading that because Qt implements COW (copy on write) for most of it is structures. Passing by const reference makes copies incredibly cheap.