documentview: fix whatsthis KUIT context marker
The translation context marker for mZoomToFitAction contains a spurious
comma, which causes xi18n to complain about an unknown subcue when
gwenview starts.

This change removes the comma, and slightly adjusts the informal part of
the context string to compensate for it being slightly harder to parse
without the comma.