Add option to show hidden files
Currently, Gwenview does not list hidden files but it's capable of opening
them from Dolphin. Once the file is open, using the arrow keys or changing
the directory will lose access to the hidden file(s).
This patch adds the feature to list hidden files following some of Dolphin's
design choices.

Like Dolphin, the `Show Hidden Files` is a toggle menu item under the `View`
menu, and is placed in the same section as `Sort By` sub-menu. Also, it uses
the standard `Show Hidden Files` shortcut (Ctrl + H).
The option is not active by default, but unlike Dolphin, it is not
persistent among instances of Gwenview, so, the user will have to activate
the option for each instance to show the hidden files in a directory.

FEATURE: 194716