Fix Gwenview being unable to open OpenRaster (ora) files
Despite having an ORA image plugin in kimageformats, Gwenview was unable to open
them. It instead preferred to open them as a zip file, which is technically
right but UX-wise wrong.

There was an explicit workaround to stop this from happening for svgz files
which are also ZIPs. We should instead check the supported image mime types
from Qt, and use that instead of hardcoding specific mimetypes.