Fix crash when crop area is smaller than handles
The crash occurs due to a failed assertion in `qBound()` that tests
`Q_ASSERT(!(max < min))`. This happens when computing the position
of the crop area middle handles (top center, bottom center, left
 center, right center) and they overlap with the corner handles.

The fix proposed is to check if there is enough room within the
respective axis (width or height) for a middle handle to be
repositioned when the crop area is not completely inside the
viewport. This preserves the existing functionality added in
commits `307816e2819f` and `e61ce1e2886c6`.

BUG: 481140