Fixing Gwenview print preview window size in multi-screen setups
By inserting a resize of the preview dialog it works around the bug
exclusive to multi-screen systems where this preview window is opened
with a tiny size.

This solution is adopted from a portion in bool KatePrinter::printPreview()

It has been suggested that this is a Qt bug. While that may be so, the
purpose of this is to provide a workable solution until the root cause
is addressed properly. So far I have not found an existing Qt issue for this.

BUG: 504978


(cherry picked from commit 2974b19e0cb5e0265cef7d932648a43162177a76)

Co-authored-by: Matthias Danzl <md@ml1.net>