Add KColorSchemeManager for Window Color Menu
While Gwenview already has the ability to change the view background
to light/medium/dark, the application chrome would always follow the
system color scheme.

This adds a KColorSchemeManager to be able to set a custom color scheme
to e.g. have Gwenview always dark regardless of the rest of the system.